refactor: move all JavaScript under ./js/ per issue #35 requirement 4.1
Issue #35 asked for "all JavaScript should live in ./js/". The previous
PR scope deferred this to a follow-up; reviewer feedback on PR #36 asked
for it to be folded back in. This commit completes the migration:

* js/src/ — application source consumed by the deployed HTML pages
  (app/, transformation/{text-to-qp-transformer,text-transformer-test}.js,
  and the nine top-level browser modules).
* js/scripts/ — Node-only CI scripts and integration runners
  (former scripts/*.mjs plus the eight repo-root *-test.mjs files).
* js/tests/ — node:test unit suites (unit/) and Playwright e2e specs
  (e2e/), preserving the template's tests/ convention one level down.

Deployed pages stay at their public URLs (app.html, entities.html,
properties.html, transformation/{index,test-ngram}.html, demo HTMLs) and
have their <script src> + import paths rewritten to point at js/src/...
The CI workflow, package.json scripts, both READMEs, the case study,
and _config.yml's exclude list are updated to match.

All gating tests still pass locally:
* node js/scripts/check-mjs-syntax.mjs    — 29 files clean
* node js/scripts/run-unit-tests.mjs      — 15/15 gating tests pass
* node js/scripts/run-e2e-local.mjs       — 9/9 Playwright tests pass
