+++ browser-cache-test.html
@@ -75,7 +75,7 @@
   </div>
 
   <script type="module">
-    import { WikidataAPIClient, WikidataSearchUtility } from './wikidata-api-browser.js';
+    import { WikidataAPIClient, WikidataSearchUtility } from './js/src/wikidata-api-browser.js';
 
     // Create global instances for testing
     window.apiClient = new WikidataAPIClient('indexeddb');
