+++ cache-demo.html
@@ -200,8 +200,8 @@
   </div>
 
   <script type="module">
-    import { TextToQPTransformer } from './transformation/text-to-qp-transformer.js';
-    import { CacheFactory } from './wikidata-api-browser.js';
+    import { TextToQPTransformer } from './js/src/transformation/text-to-qp-transformer.js';
+    import { CacheFactory } from './js/src/wikidata-api-browser.js';
 
     let currentTransformer = new TextToQPTransformer();
     let currentCacheType = 'auto';
