+++ js/src/app/modes/alphabet.jsx
