+++ js/src/app/tests-panel.jsx
