+++ js/src/wikidata-api-browser.js
