+++ run-tests.html
@@ -97,7 +97,7 @@
   </div>
 
   <script type="module">
-    import { TextTransformerTest, demonstrateTransformer } from './transformation/text-transformer-test.js';
+    import { TextTransformerTest, demonstrateTransformer } from './js/src/transformation/text-transformer-test.js';
 
     let testSuite;
     let output = document.getElementById('output');
