+++ search-demo.html
@@ -311,8 +311,8 @@
     </div>
 
     <script type="module">
-        import { searchUtility } from './wikidata-api-browser.js';
-        import { WikidataSearchTest, demonstrateSearchFeatures } from './search-test.js';
+        import { searchUtility } from './js/src/wikidata-api-browser.js';
+        import { WikidataSearchTest, demonstrateSearchFeatures } from './js/src/search-test.js';
 
         // Make functions globally available
         window.searchUtility = searchUtility;
