+++ transformation/test-ngram.html
@@ -61,7 +61,7 @@
   <div id="results"></div>
 
   <script type="module">
-    import { TextToQPTransformer } from './text-to-qp-transformer.js';
+    import { TextToQPTransformer } from '../js/src/transformation/text-to-qp-transformer.js';
 
     window.runTests = async function() {
       const results = document.getElementById('results');
